requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
pymongo==3.12.0
python-dotenv==0.21.0
gunicorn==20.1.0
